chore: rename functions
